add some API Codes (setmemberCoins, getMemberInfos, changeCurrentBorder, getMemberAllBorders, byuBorder)

+++ prototype/src/main/java/com/klab2/challenge/prototype/domain/MemberInfos.java
@@ -13,6 +13,6 @@ import lombok.NoArgsConstructor;
 public class MemberInfos {
     private int totalCoins;
     private int holdingCoins;
-    private int currentBorder;
+    private Long currentBorder;
     private String imageUrl;
 }
